Refactor team role checks and add unit tests
Updated authorization logic in ArbeitsgruppeRequest, FanfictionRequest, and KassenbuchEntryRequest to check roles explicitly in the members team instead of using currentTeam. Improved documentation in MembersTeamAware and CreatesUserWithRole traits. Changed calculatePaymentAmount return type to float in FantreffenRegistrationService. Added unit tests for FantreffenRegistrationService.

+++ app/Services/FantreffenRegistrationService.php
@@ -65,9 +65,9 @@ class FantreffenRegistrationService
      * @param  bool  $tshirtBestellt  Ob ein T-Shirt bestellt wurde
      * @param  bool  $isAuthenticated  Ob der User eingeloggt (Mitglied) ist
      */
-    public function calculatePaymentAmount(bool $tshirtBestellt, bool $isAuthenticated): int
+    public function calculatePaymentAmount(bool $tshirtBestellt, bool $isAuthenticated): float
     {
-        $amount = 0;
+        $amount = 0.0;
 
         // Gäste zahlen Grundgebühr
         if (! $isAuthenticated) {
